+++ src/model/characters/player/clicker/ClickerTools.java
@@ -1,7 +1,7 @@
 package model.characters.player.clicker;
 
 public class ClickerTools{
-    protected double resourceAmount = 1;
+    protected int resourceAmount = 1;
     public ClickerTools(){
     }
     public double getResourceAmount(){
